refactor(core): consolidate projections into algebra
Epic 1 - The Monolith Consolidation:
- Migrated all functions and variables from `src/coreason_manifest/utils/projections.py` to `src/coreason_manifest/utils/algebra.py`.
- Topologically sorted imports in `algebra.py` (stdlib, third-party, local).
- Deleted `src/coreason_manifest/utils/projections.py` to eliminate directory fragmentation.
- Rewrote `src/coreason_manifest/__init__.py` to import defederated logic from `algebra.py`.
- Updated test dependencies in `tests/contracts/test_projections.py`.
